fix: Add route handler for Next Auth, include user session in the root layout
