+++ client/app/layout.jsx
@@ -1,16 +1,13 @@
-import { SessionProvider } from "next-auth/react";
+import SessionProvider from "./_components/SessionProvider";
+import { getServerSession } from "next-auth";
 
-export default function Layout({ children, session }) {
+export default async function Layout({ children }) {
+  const session = await getServerSession();
   return (
     <html>
       <body>
-        <SessionProvider session={session}>
-          {children}
-        </SessionProvider>
+        <SessionProvider session={session}>{children}</SessionProvider>
       </body>
     </html>
   );
 }
-
-
-
